+++ devtools/replidev/src/command/play/mod.rs
@@ -4,11 +4,16 @@ use crate::conf::Conf;
 use crate::ErrorKind;
 use crate::Result;
 
+mod node_list;
 mod node_start;
 
 /// Manage Replicante Playground nodes.
 #[derive(Debug, StructOpt)]
 pub enum CliOpt {
+    /// List all playground nodes.
+    #[structopt(name = "node-list")]
+    NodeList,
+
     /// Start a new playground node.
     #[structopt(name = "node-start")]
     NodeStart(NodeOpt),
@@ -16,6 +21,10 @@ pub enum CliOpt {
 
 #[derive(Debug, StructOpt)]
 pub struct NodeOpt {
+    /// ID of the cluster to place the node into.
+    #[structopt(name = "cluster-id", long)]
+    cluster_id: Option<String>,
+
     /// Store node to start.
     #[structopt(name = "STORE", required = true)]
     store: String,
@@ -28,6 +37,7 @@ pub fn run(args: CliOpt, conf: Conf) -> Result<bool> {
         return Err(error.into());
     }
     match args {
+        CliOpt::NodeList => node_list::run(&conf),
         CliOpt::NodeStart(start) => node_start::run(&start, &conf),
     }
 }