+++ devtools/replidev/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - HTTPS certificate (re)generation command.
 - Initialise and clean up replicante core dependencies.
 - List available `replidev deps` pods.
+- List running playground nodes.
 - Optional per-machine configuration override.
 - Start and stop replicante core dependencies pods.
 - Start playground nodes as pods.